Create a notesThunk with createJobApplicationNote and getAllJobApplicationNotes
